[新增]若预览的视频为H265格式，则返回“不支持的视频格式”.

+++ projects/HDview/src/com/starnet/snview/protocol/message/OwspVideoDataFormat.java
@@ -1,7 +1,7 @@
 package com.starnet.snview.protocol.message;
 
 public class OwspVideoDataFormat {
-	private int codecId;
+	private long codecId;
 	private int bitrate;
 	private int width;
 	private int height;
@@ -10,10 +10,10 @@ public class OwspVideoDataFormat {
 	private int reserve;
 	
 	
-	public int getCodecId() {
+	public long getCodecId() {
 		return codecId;
 	}
-	public void setCodecId(int codecId) {
+	public void setCodecId(long codecId) {
 		this.codecId = codecId;
 	}
 	public int getBitrate() {
